knowledge_base/requirements.txt
# ============================================================
# NSEQA - 神经-符号混合分布式知识专家问答系统 依赖
# ============================================================

# OpenAI API 客户端（所有模型调用统一走 OpenAI 兼容接口）
openai>=1.0.0

# 向量检索
faiss-cpu>=1.7.4  # GPU 版本使用 faiss-gpu
numpy>=1.24.0

# 图处理
networkx>=3.0
python-louvain>=0.16

# NLP 工具
spacy>=3.5.0
nltk>=3.8

# 数据处理
pandas>=2.0.0

# 主题模型 / 聚类（用于专家领域自动识别）
gensim>=4.3.0
scikit-learn>=1.2.0

# 配置管理
pyyaml>=6.0

# 进度条
tqdm>=4.65.0

# 测试
pytest>=7.3.0
